+++ lib/server/routes.js
@@ -1,7 +1,7 @@
 "use strict";
 
 const config = require("exp-config");
-const {getMetrics} = require("./metrics");
+const {getMetrics} = require("../metrics");
 const {logger} = require("lu-logger");
 const triggerHandler = require("./trigger-route");
