+++ lib/server/trigger-route.js
@@ -1,5 +1,7 @@
 "use strict";
-const {crd} = require("./broker");
+const {crd} = require("../broker");
+const parseBody = require("./parse-body");
+
 function triggerHandler(key) {
   return async (req, res) => {
     const {correlationId} = req;
@@ -22,32 +24,6 @@ function triggerHandler(key) {
     res.end();
   };
 }
-
-function parseBody(req) {
-  return new Promise((resolve, reject) => {
-    const chunks = [];
-    req.on("data", (chunk) => {
-      chunks.push(chunk);
-    });
-    req.on("end", () => {
-      try {
-        const body = chunks.length > 0 ? JSON.parse(Buffer.concat(chunks).toString()) : {};
-        resolve(body);
-      } catch (err) {
-        reject(err);
-      }
-    });
-    req.on("error", reject);
-  });
-}
-
-function renderError(errors, res, req) {
-  const {correlationId} = req;
-  res.writeHead(400, {"Content-Type": "application/json", "correlation-id": correlationId});
-  res.write(JSON.stringify({errors: errors.map(buildError), meta: {correlationId}}));
-  res.end();
-}
-
 function validate(source) {
   const errors = [];
   for (const field of ["type", "id", "attributes"]) {
@@ -60,6 +36,13 @@ function validate(source) {
   }
 
   return errors;
+}
+
+function renderError(errors, res, req) {
+  const {correlationId} = req;
+  res.writeHead(400, {"Content-Type": "application/json", "correlation-id": correlationId});
+  res.write(JSON.stringify({errors: errors.map(buildError), meta: {correlationId}}));
+  res.end();
 }
 
 function buildError(detail) {
